Move build_schema_test to the restapi testtools module
